Remove builder, consolidate properties/display.rs
