+++ examples/graphics.rs
@@ -29,7 +29,7 @@ use embedded_graphics::{
     primitives::{Circle, Line, Rectangle},
 };
 use panic_semihosting as _;
-use ssd1331::Builder;
+use ssd1331::{DisplayRotation::Rotate0, Ssd1331};
 use stm32f1xx_hal::delay::Delay;
 use stm32f1xx_hal::prelude::*;
 use stm32f1xx_hal::spi::{Mode, Phase, Polarity, Spi};
@@ -73,7 +73,7 @@ fn main() -> ! {
         &mut rcc.apb2,
     );
 
-    let mut disp = Builder::new().connect_spi(spi, dc);
+    let mut disp = Ssd1331::new(spi, dc, Rotate0);
 
     disp.reset(&mut rst, &mut delay).unwrap();
     disp.init().unwrap();
