+++ examples/rotation.rs
@@ -29,7 +29,7 @@ use cortex_m_rt::ExceptionFrame;
 use cortex_m_rt::{entry, exception};
 use embedded_graphics::prelude::*;
 use panic_semihosting as _;
-use ssd1331::{Builder, DisplayRotation};
+use ssd1331::{DisplayRotation, Ssd1331};
 use stm32f1xx_hal::delay::Delay;
 use stm32f1xx_hal::prelude::*;
 use stm32f1xx_hal::spi::{Mode, Phase, Polarity, Spi};
@@ -73,17 +73,14 @@ fn main() -> ! {
         &mut rcc.apb2,
     );
 
-    let mut disp = Builder::new()
-        // Set initial rotation at 90 degrees clockwise
-        .rotation(DisplayRotation::Rotate90)
-        .connect_spi(spi, dc);
+    // Initialise the display with a default rotation of 90 degrees
+    let mut disp = Ssd1331::new(spi, dc, DisplayRotation::Rotate90);
 
     disp.reset(&mut rst, &mut delay).unwrap();
     disp.init().unwrap();
     disp.flush().unwrap();
 
-    // Contrived example to test builder and instance methods. Sets rotation to 270 degress
-    // or 90 degress counterclockwise
+    // Set a new rotation of 270 degrees
     disp.set_rotation(DisplayRotation::Rotate270).unwrap();
 
     // let (w, h) = disp.get_dimensions();
